Docker_Project1/requirements.txt
flask==2.3.3
flask_sqlalchemy
psycopg2-binary
redis
python-dotenv